# back-end/swagger.yaml
openapi: 3.0.3
info:
  title: APIs tisco - OpenAPI 3.0
  description: APIs for tisco
  version: 1.0.0
servers:
  - url: http://localhost:3000/api/v1
  - url: http://159.138.242.188:3000/api/v1
...
components:
  securitySchemes:
    beareAuth:
        type: http
        scheme: bearer
        name: Authorization
        in: header
        bearerFormat: JWT
        description: >-
          Enter the token with the `Bearer: ` prefix, e.g. "Bearer abcde12345"
...
paths:
...
  /consent/get-consent:
    get:
      tags:
        - consent
      security:
        - beareAuth: []
      summary: Get consent
      operationId: getConsent
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  consent:
                    type: object
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /consent/save-consent:
    post:
      tags:
        - consent
      security:
        - beareAuth: []
      summary: Save consent
      operationId: saveConsent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                authen:
                  type: object
                  properties:
                    uuid:
                      type: string
                option:
                  type: integer
                # Add other properties as needed
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  save:
                    type: boolean
                  saveMember:
                    type: object
                  verifyMember:
                    type: object
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
...
  /game/categories:
    get:
      tags:
        - game-board
      security:
        - beareAuth: []
      summary: Get categories
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  categories:
                    type: array
                    items:
                      type: string
  /game/board-by-cate:
    get:
      tags:
        - game-board
      security:
        - beareAuth: []
      summary: Get board game by categories
      parameters:
        - in: path
          name: ids
          description: Comma-separated list of category IDs
          schema:
            type: string
      responses:
        '200':
          description: Success
  /game/board/{id}:
    get:
      tags:
        - game-board
      security:
        - beareAuth: []
      summary: Get game board
      parameters:
        - in: path
          name: id
          description: ID of the game board
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Success
  /game/save-answer:
    post:
      tags:
        - game-board
      security:
        - beareAuth: []
      summary: Save game board answer
      operationId: createHistory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                gmBoardId: 
                  type: number
                gmGameLevel:
                  type: number
                gmScore:  
                  type: number
                gmHisSystem:
                  type: object
                gmHisRecord:
                  type: array
                  items:
                    type: object
                    properties:
                      question:
                        type: string
                      answers:
                        type: array
                        items:
                          type: object
                          properties:
                            answer:
                              type: string
                            ansScore:
                              type: integer
                    required:
                    - question
                    - answers
              required:
                - gmBoardId
                - gmGameLevel
                - gmScore
                - gmHisSystem
                - gmHisRecord
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string